+++ tests/test_tcod.py
@@ -101,62 +101,6 @@ def test_tcod_map_pickle_fortran():
     assert map_.fov.strides == map2.fov.strides
 
 
-@pytest.mark.parametrize('implementation', [tcod.noise.SIMPLE,
-                                            tcod.noise.FBM,
-                                            tcod.noise.TURBULENCE])
-def test_noise_class(implementation):
-    noise = tcod.noise.Noise(2, tcod.NOISE_SIMPLEX, implementation)
-    # cover attributes
-    assert noise.dimensions == 2
-    noise.algorithm = noise.algorithm
-    noise.implementation = noise.implementation
-    noise.octaves = noise.octaves
-    assert noise.hurst
-    assert noise.lacunarity
-
-    noise.get_point(0, 0)
-    noise.sample_mgrid(np.mgrid[:2,:3])
-    noise.sample_ogrid(np.ogrid[:2,:3])
-
-
-def test_noise_samples():
-    noise = tcod.noise.Noise(2, tcod.NOISE_SIMPLEX, tcod.noise.SIMPLE)
-    np.testing.assert_equal(
-        noise.sample_mgrid(np.mgrid[:32,:24]),
-        noise.sample_ogrid(np.ogrid[:32,:24]),
-        )
-
-
-def test_noise_errors():
-    with pytest.raises(ValueError):
-        tcod.noise.Noise(0)
-    with pytest.raises(ValueError):
-        tcod.noise.Noise(1, implementation=-1)
-    noise = tcod.noise.Noise(2)
-    with pytest.raises(ValueError):
-        noise.sample_mgrid(np.mgrid[:2,:2,:2])
-    with pytest.raises(ValueError):
-        noise.sample_ogrid(np.ogrid[:2,:2,:2])
-
-
-@pytest.mark.parametrize('implementation',
-    [tcod.noise.SIMPLE, tcod.noise.FBM, tcod.noise.TURBULENCE])
-def test_noise_pickle(implementation):
-    rand = tcod.random.Random(tcod.random.MERSENNE_TWISTER, 42)
-    noise = tcod.noise.Noise(2, implementation, seed=rand)
-    noise2 = copy.copy(noise)
-    assert (noise.sample_ogrid(np.ogrid[:3,:1]) ==
-            noise2.sample_ogrid(np.ogrid[:3,:1])).all()
-
-
-def test_noise_copy():
-    rand = tcod.random.Random(tcod.random.MERSENNE_TWISTER, 42)
-    noise = tcod.noise.Noise(2, seed=rand)
-    noise2 = pickle.loads(pickle.dumps(noise))
-    assert (noise.sample_ogrid(np.ogrid[:3,:1]) ==
-            noise2.sample_ogrid(np.ogrid[:3,:1])).all()
-
-
 @pytest.mark.filterwarnings("ignore")
 def test_color_class():
     assert tcod.black == tcod.black